advertised network isolation: move advertised network isolation to udn_isolation.go
